gpt: Revert small device commits
The following commit broke the position of the LastUsableLBA:

48f236f9: gpt: permit "mklabel gpt" on a 67-sector device

It introduced an off by one error, leaving LastUsableLBA pointing to
the first sector of the backup partition table instead.

This effectively reverts that commit, as well as adjusting the subsequent
commits to instead use 68 sectors as the minimum length.  I believe
this is the minimum legal size as at 67 sectors, there is no valid
value for FirstUsableLBA and LastUsableLBA.

Conflicts:
	tests/t0203-gpt-create-on-min-sized-device.sh